agents/political-news/pyproject.toml
[tool.poetry]
name = "political-news"
version = "0.1.0"
description = "Political News Agent: Scrapes unbiased political news from the last 24 hours"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "political_news"}]

[tool.poetry.dependencies]
python = "^3.9"
google-adk = "^0.1.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
deprecated = "^1.2.18"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 